fix: NavBar 위치 및 코드 수정
